% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_time.R
\name{Alfaifi}
\alias{Alfaifi}
\alias{prep_time}
\title{Calculate Prep Time}
\usage{
prep_time(.ingredients_choice)
}
\arguments{
\item{The}{ingredients chosen by the use an produced by the funtion prepare_data}
}
\value{
The Prep Time
}
\description{
Calculate the preparation Time
}
